renamed normal button to dumbButton
