+++ src/components/Button/index.ts
@@ -1,2 +1,2 @@
-export { default as Button } from './component'
+export { default as DumbButton } from './DumbButton'
 export { default as AsyncButton } from './AsyncButton'
